New version 301: various new effects, changes to sienaRI, some smaller changes.
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rsiena@301 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/RSienaTest/src/model/effects/FourCyclesEffect.h
@@ -52,12 +52,12 @@ private:
 	// Indicates if the effect is used for a two-mode or one-mode network
 	bool lTwoMode;
 
-	void countThreePaths(int i, const Network * pNetwork, int * counters)
+	void countThreePaths(int i, const Network * pNetwork, long int * counters)
 		const;
 	// For a fixed i, this variable stores the number of three-paths
 	// i -> h <- k -> j per each j.
 
-	int * lcounters;
+	long int * lcounters;
 
 	// Indicates if the square root of the number of four-cycles has to
 	// be taken.
